Avoid raising an exception in an at_exit handler
Before this patch linking mirage-unix on Windows would register the
at_exit handler, which would attempt to ignore SIGPIPE on exit, which
would raise an exception because the signal is not supported by the
platform.

If a platform does not support SIGPIPE, it does not matter if attempts
to ignore the exception fail.

Signed-off-by: David Scott <dave@recoil.org>